Time: 5 ms (50.00%), Space: 52.9 MB (100.00%) - LeetHub

+++ 2875-minimum-size-subarray-in-infinite-array/2875-minimum-size-subarray-in-infinite-array.java
@@ -5,7 +5,7 @@ class Solution {
         for(int i = 0; i < n; i++)
             tot += nums[i];
         int ans = 0;
-        if(target > (long)tot)
+        if(target > tot)
         {
             ans = (int)(target/tot) * n;
             target = (int)(target % tot);
